Added aspects of the about section
The profile picture and the resume buttons have been added for the about-me section. The animations are causing the CSS to get a bit messy, so I might move it all over to JS. The buttons are not finished since the text that appears below them is not displaying properly. A "welcome" message was also added to the top of index.html for anyone who views the source code.