Introduced a functioning block animation
The block animation now works however it varies slightly from the specified design. I determined to effectively generate blocks, I needed to select where they generated as to keep integrity of the backgroun and not leave it all to random chance. This required me to introduce a less flexible design in favor of more control. I believe this trade-off was worth it.  Slight changes were made to the wave animation to accomodate the block animation.